@@ -15,7 +15,6 @@
   "scripts": {
     "start": "npm run dev",
     "dev": "next",
-    "prebuild": "contentlayer build",
     "build": "next build",
     "postbuild": "next-sitemap --config next-sitemap.config.js",
     "serve": "next start",
@@ -26,29 +25,25 @@
   },
   "dependencies": {
     "@docsearch/react": "^3.5.2",
-    "@next/mdx": "^13.5.0",
-    "contentlayer": "^0.3.4",
+    "@mdx-js/loader": "^3.0.0",
+    "@mdx-js/react": "^3.0.0",
+    "@next/mdx": "^14.0.4",
+    "@tabler/icons-react": "^2.5.0",
+    "@types/mdx": "^2.0.10",
     "framer-motion": "^10.16.4",
     "globby": "^13.2.2",
-    "next-contentlayer": "^0.3.4",
-    "prism-react-renderer": "^1.3.5",
-    "@tabler/icons-react": "^2.5.0",
-    "next": "^13.5.4",
+    "next": "^14.0.0",
+    "next-mdx-remote": "^4.4.1",
     "next-pwa": "^5.6.0",
-    "next-seo": "^6.1.0",
     "next-sitemap": "^4.2.3",
     "next-themes": "^0.2.1",
+    "prism-react-renderer": "^1.3.5",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   },
   "devDependencies": {
+    "@ianvs/prettier-plugin-sort-imports": "^4.1.1",
     "@tailwindcss/typography": "^0.5.10",
-    "reading-time": "^1.5.0",
-    "rehype-code-titles": "^1.1.0",
-    "rehype-slug": "^5.1.0",
-    "rehype-toc": "^3.0.2",
-    "remark-gfm": "^3.0.1",
-    "remark-sectionize": "^1.1.1",
     "@types/node": "^20.8.2",
     "@types/react": "18.2.25",
     "@types/react-dom": "^18.2.10",
@@ -68,13 +63,32 @@
     "postcss-import": "^15.1.0",
     "prettier": "^3.0.3",
     "prettier-plugin-tailwindcss": "^0.5.5",
+    "reading-time": "^1.5.0",
+    "rehype-code-titles": "^1.1.0",
+    "rehype-slug": "^5.1.0",
+    "rehype-toc": "^3.0.2",
+    "remark-gfm": "^3.0.1",
+    "remark-sectionize": "^2.0.0",
     "tailwindcss": "^3.3.3",
     "typescript": "^5.2.2"
   },
   "prettier": {
     "singleQuote": true,
     "semi": false,
-    "printWidth": 120
+    "printWidth": 120,
+    "plugins": [
+      "@ianvs/prettier-plugin-sort-imports",
+      "prettier-plugin-tailwindcss"
+    ],
+    "importOrder": [
+      "^react$",
+      "",
+      "<THIRD_PARTY_MODULES>",
+      "",
+      "^@/(.*)$",
+      "",
+      "^[.]"
+    ]
   },
   "eslintConfig": {
     "extends": [
@@ -89,6 +103,12 @@
         "error",
         {
           "ignoreRestSiblings": true
+        }
+      ],
+      "@typescript-eslint/no-explicit-any": [
+        "error",
+        {
+          "ignoreRestArgs": true
         }
       ]
     }
