@@ -1,10 +1,14 @@
-import NextLink from 'next/link'
-import { useRouter } from 'next/router'
+'use client'
+
 import React, { useEffect, useMemo, useState } from 'react'
 
+import NextLink from 'next/link'
+import { usePathname } from 'next/navigation'
+
 export default function Link({ href, children, className, ...props }: React.ComponentProps<'a'>) {
-  const { asPath } = useRouter()
   const [origin, setOrigin] = useState<string | null>(null)
+
+  const pathname = usePathname()
 
   useEffect(() => {
     setOrigin(window.location.origin)
@@ -15,8 +19,8 @@ export default function Link({ href, children, className, ...props }: React.Comp
       href === undefined ||
       href.startsWith('#') ||
       href.startsWith('?') ||
-      (origin !== null && new URL(asPath, origin).pathname === new URL(href, origin).pathname),
-    [href, origin, asPath]
+      (origin !== null && new URL(pathname, origin).pathname === new URL(href, origin).pathname),
+    [href, origin, pathname],
   )
 
   if (href === undefined)
@@ -31,7 +35,7 @@ export default function Link({ href, children, className, ...props }: React.Comp
       {children}
     </a>
   ) : (
-    <NextLink href={href} className={className}>
+    <NextLink href={href} className={`text-brand ${className}`}>
       {children}
     </NextLink>
   )
