@@ -1,5 +1,3 @@
-const { withContentlayer } = require('next-contentlayer')
-
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   output: 'export',
@@ -13,6 +11,8 @@ const nextConfig = {
   },
 }
 
+const withMDX = require('@next/mdx')()
+
 const withPWA = require('next-pwa')({
   dest: 'public',
   disable: process.env.NODE_ENV === 'development',
@@ -20,4 +20,4 @@ const withPWA = require('next-pwa')({
   sw: 'service-worker.js',
 })
 
-module.exports = withPWA(withContentlayer(nextConfig))
+module.exports = withPWA(withMDX(nextConfig))
