@@ -1,6 +1,9 @@
+'use client'
+
+import React, { useEffect, useRef } from 'react'
+
 import CodeBlock, { CodeBlockProps } from '@/components/CodeBlock'
 import Window, { WindowProps } from '@/components/Window'
-import React, { useEffect, useRef } from 'react'
 
 export interface CodeWindowProps extends WindowProps, CodeBlockProps {
   autoScroll?: boolean
