@@ -1,12 +1,16 @@
-import { SearchButton } from '@/components/Search'
-import { IconSearch } from '@tabler/icons-react'
+'use client'
+
 import React from 'react'
+
+import { IconSearch } from '@tabler/icons-react'
+
+import { SearchButton } from '@/components/Search'
 
 export function QuickSearchButton({ className, ...props }: React.ComponentProps<'button'>) {
   return (
     <SearchButton
       {...props}
-      className={`dark:highlight-white/5 flex flex-1 items-center gap-x-3 rounded-full bg-white px-4 text-left shadow-sm ring-1 ring-brand-900/10 hover:ring-brand-300 focus:outline-none focus:ring-2 focus:ring-brand-500 dark:bg-primary-700 dark:ring-0 dark:hover:bg-primary-600 ${className}`}
+      className={`dark:highlight-white/5 flex flex-1 items-center gap-x-3 rounded-full bg-white px-4 text-left shadow-sm ring-1 ring-brand-900/10 transition-all duration-200 hover:ring-brand-300 focus:outline-none focus:ring-2 focus:ring-brand-500 dark:bg-primary-700 dark:ring-0 dark:hover:bg-primary-600 ${className}`}
     >
       {({ actionKey }) => (
         <>
