@@ -1,7 +1,9 @@
+import React from 'react'
+
 // eslint-disable-next-line import/named
 import { DocSearchHit } from '@docsearch/react/dist/esm/types'
-import Link from 'next/link'
-import React from 'react'
+
+import Link from '@/components/Link'
 
 interface IndexedDocSearchHit extends DocSearchHit {
   __is_result(): boolean
