@@ -1,6 +1,9 @@
+'use client'
+
+import React, { useContext, useEffect, useRef } from 'react'
+
 import { SearchContext } from '@/components/Search/Context'
 import { ActionKey, useActionKey } from '@/hooks/useActionKey'
-import React, { useContext, useEffect, useRef } from 'react'
 
 interface ChildrenProps {
   actionKey: ActionKey
