@@ -1,6 +1,7 @@
-import { IconMoon, IconSun } from '@tabler/icons-react'
-import { useTheme } from 'next-themes'
 import React, { useCallback, useEffect, useState } from 'react'
+
+import { IconBrushOff, IconMoon, IconSun } from '@tabler/icons-react'
+import { useTheme } from 'next-themes'
 
 export default function ThemeModeSwitcher({ ...props }: React.ComponentProps<'button'>) {
   const [mounted, setMounted] = useState(false)
@@ -12,11 +13,17 @@ export default function ThemeModeSwitcher({ ...props }: React.ComponentProps<'bu
 
   useEffect(() => setMounted(true), [])
 
-  if (!mounted) return null
-
   return (
     <button aria-label="Theme Mode Switcher" type="button" {...props} onClick={onSwitch}>
-      {theme === 'light' ? <IconSun className="h-full w-full" /> : <IconMoon className="h-full w-full" />}
+      {mounted ? (
+        theme === 'light' ? (
+          <IconSun className="h-full w-full" />
+        ) : (
+          <IconMoon className="h-full w-full" />
+        )
+      ) : (
+        <IconBrushOff className="h-full w-full" />
+      )}
     </button>
   )
 }
