@@ -1,5 +1,6 @@
-import { IconCircleMinus, IconCirclePlus, IconCircleX } from '@tabler/icons-react'
 import React, { MutableRefObject, useCallback, useState } from 'react'
+
+import { IconCircleMinus, IconCirclePlus, IconCircleX } from '@tabler/icons-react'
 
 export interface WindowProps extends React.ComponentProps<'div'> {
   title?: string
@@ -32,8 +33,8 @@ export default function Window({ title, contentRef, className, children }: Windo
           state === 'open' ? 'relative dark:bg-primary-950/70' : ''
         } ${className || ''}`}
       >
-        <div className="mb-px flex h-8 w-full items-center justify-between border-b border-primary-500 px-4 text-primary-400 dark:border-primary-500/30 dark:text-primary-500">
-          <span className="truncate font-semibold text-primary-400">{title}</span>
+        <div className="mb-px flex h-8 w-full items-center justify-between border-b border-primary-500 px-4 text-primary-300 dark:border-primary-500/30 dark:text-primary-400">
+          <span className="truncate font-semibold">{title}</span>
           <div className="flex items-center justify-end gap-x-2">
             <button className="h-4 w-4" onClick={onMinimize} aria-label="Minimize Window">
               <IconCircleMinus className="h-full w-full" />
