@@ -1,5 +1,5 @@
 import React from 'react'
 
 export default function BosqueName() {
-  return <span className="text-bosque-600 dark:text-bosque-400">Bosque</span>
+  return <span className="text-bosque-500">Bosque</span>
 }
