@@ -1,5 +1,5 @@
 import React from 'react'
 
 export default function BrumaName() {
-  return <span className="text-bruma-600 dark:text-bruma-400">Bruma</span>
+  return <span className="text-bruma-500">Bruma</span>
 }
