@@ -1,5 +1,5 @@
 import React from 'react'
 
 export default function CiclonName() {
-  return <span className="text-ciclon-600 dark:text-ciclon-400">Ciclon</span>
+  return <span className="text-ciclon-500">Ciclon</span>
 }
