@@ -1,5 +1,5 @@
 import React from 'react'
 
 export default function FlamaName() {
-  return <span className="text-flama-600 dark:text-flama-400">Flama</span>
+  return <span className="text-flama-500">Flama</span>
 }
