@@ -1,5 +1,5 @@
 import React from 'react'
 
 export default function PythonName() {
-  return <span className="text-green-600 dark:text-green-400">Python</span>
+  return <span className="text-green-500">Python</span>
 }
