@@ -1,5 +1,5 @@
 import React from 'react'
 
 export default function VorticoName() {
-  return <span className="text-vortico-600 dark:text-vortico-400">Vortico</span>
+  return <span className="text-vortico-500">Vortico</span>
 }
